// src/types/IElement.ts
export enum EAllergen {
  CELERY = 'celery',
  CORN = 'corn',
  CRUSTACEAN_SHELLFISH = 'crustacean_shellfish',
  EGG = 'egg',
  FISH = 'fish',
  FRUITS = 'fruits',
  GELATIN = 'gelatin',
  GLUTEN = 'gluten',
  LUPIN = 'lupin',
  MEAT = 'meat',
  MILK = 'milk',
  MOLLUSKS = 'mollusks',
  MUSTARD = 'mustard',
  PEANUTS = 'peanuts',
  SESAME = 'sesame',
  SOYBEANS = 'soybeans',
  SPICES = 'spices',
  TREE_NUTS = 'tree_nuts',
  WHEAT = 'wheat',
}

export enum EVitamin {
  VITAMIN_A = 'vitamin_a',
  VITAMIN_B1 = 'vitamin_b1',
  VITAMIN_B12 = 'vitamin_b12',
  VITAMIN_B2 = 'vitamin_b2',
  VITAMIN_B3 = 'vitamin_b3',
  VITAMIN_B5 = 'vitamin_b5',
  VITAMIN_B6 = 'vitamin_b6',
  VITAMIN_B7 = 'vitamin_b7',
  VITAMIN_B9 = 'vitamin_b9',
  VITAMIN_C = 'vitamin_c',
  VITAMIN_D = 'vitamin_d',
  VITAMIN_E = 'vitamin_e',
  VITAMIN_K = 'vitamin_k',
}

export enum EMineral {
  CALCIUM = 'calcium',
  CHLORIDE = 'chloride',
  CHROMIUM = 'chromium',
  COBALT = 'cobalt',
  COPPER = 'copper',
  FLUORIDE = 'fluoride',
  IODINE = 'iodine',
  IRON = 'iron',
  MAGNESIUM = 'magnesium',
  MANGANESE = 'manganese',
  MOLYBDENUM = 'molybdenum',
  PHOSPHORUS = 'phosphorus',
  POTASSIUM = 'potassium',
  SELENIUM = 'selenium',
  SODIUM = 'sodium',
  SULFUR = 'sulfur',
  ZINC = 'zinc',
}